+++ CHANGELOG.md
@@ -12,12 +12,13 @@
 
 ### Bug fixes
 
-- storage: handle errors in batch objects delete action #627
+- Storage: handle errors in batch objects delete action #627
+- Instance: Fix instance protection flag update zone context #648
 
 ## 1.80.0
 
 ### Features
-- instance pool: added min-available flag to exo compute #629
+- Instance pool: added min-available flag to exo compute #629
 - dbaas: external endpoints and external integration commands and sub-commands
 
 ## 1.79.1